requirements.txt
beautifulsoup4==4.12.3
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
DateTime==5.5
deepl==1.18.0
google==3.0.0
google-api-core==2.19.1
google-auth==2.33.0
google-cloud==0.34.0
google-cloud-core==2.4.1
google-cloud-translate==3.16.0
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.1
grpcio==1.65.4
grpcio-status==1.65.4
idna==3.7
logging==0.4.9.6
lxml==5.3.0
modernmt==1.5.2
numpy==2.0.1
pandas==2.2.2
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rsa==4.9
setuptools==72.2.0
six==1.16.0
soupsieve==2.6
tzdata==2024.1
urllib3==2.2.2
zope.interface==7.0.1
